Tweak tree placement
Since Natura is not using the DecorateBiomeEvent for whatever reason, Natura's trees are generated after the biomes are finished decorating.
This commit deals with the placement of trees, effectively only allowing to replace air in most cases, reducing conjoined trees by a lot.
This should be reworked as a whole eventually.